refactor promiseInsertUser into dbPromises helper
